+++ sdk/resourcemanager/azure-resourcemanager-appservice/src/main/java/com/azure/resourcemanager/appservice/models/SupportsOneDeploy.java
@@ -110,4 +110,20 @@ public interface SupportsOneDeploy {
      * if {@link DeployOptions#withTrackDeployment(Boolean)} set to True.
      */
     Mono<KuduDeploymentResult> pushDeployAsync(DeployType type, File file, DeployOptions deployOptions);
+
+    /**
+     * Gets the production deployment status for Azure Site.
+     *
+     * @param deploymentId GUID of the deployment operation
+     * @return the production deployment status for Azure Site
+     */
+    DeploymentBuildStatus getDeploymentStatus(String deploymentId);
+
+    /**
+     * Gets the production deployment status for Azure Site.
+     *
+     * @param deploymentId GUID of the deployment operation
+     * @return the production deployment status for Azure Site
+     */
+    Mono<DeploymentBuildStatus> getDeploymentStatusAsync(String deploymentId);
 }